+++ projects/js-packages/social-previews/src/instagram-preview/constants.tsx
@@ -1,1 +1,1 @@
-export const FEED_TEXT_MAX_LENGTH = 520;
+export const FEED_TEXT_MAX_LENGTH = 2200;