+++ projects/js-packages/analytics/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### This is a list detailing changes for the Jetpack RNA Analytics package releases.
 
+## [1.0.11] - 2026-05-04
+### Changed
+- Internal: No longer require automattic/jetpack-changelogger as a per-project dev dependency. [#48225]
+
 ## [1.0.10] - 2026-04-27
 ### Fixed
 - Include `blog_id` automatically in all Tracks events via the `jpTracksContext` global. [#48096]
@@ -192,6 +196,7 @@
 ### Added
 - Initial release of jetpack-api package.
 
+[1.0.11]: https://github.com/Automattic/jetpack-analytics/compare/v1.0.10...v1.0.11
 [1.0.10]: https://github.com/Automattic/jetpack-analytics/compare/v1.0.9...v1.0.10
 [1.0.9]: https://github.com/Automattic/jetpack-analytics/compare/v1.0.8...v1.0.9
 [1.0.8]: https://github.com/Automattic/jetpack-analytics/compare/v1.0.7...v1.0.8