+++ projects/js-packages/number-formatters/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.7] - 2026-05-04
+### Changed
+- Internal: No longer require automattic/jetpack-changelogger as a per-project dev dependency. [#48225]
+
 ## [1.1.6] - 2026-04-15
 ### Changed
 - Internal updates.
@@ -136,6 +140,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Initial release
 - Basic number formatting functionality
 
+[1.1.7]: https://github.com/Automattic/number-formatters/compare/1.1.6...1.1.7
 [1.1.6]: https://github.com/Automattic/number-formatters/compare/1.1.5...1.1.6
 [1.1.5]: https://github.com/Automattic/number-formatters/compare/1.1.4...1.1.5
 [1.1.4]: https://github.com/Automattic/number-formatters/compare/1.1.3...1.1.4