+++ projects/packages/boost-core/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.4] - 2026-05-04
+### Changed
+- Internal: No longer require automattic/jetpack-changelogger as a per-project dev dependency. [#48225]
+
 ## [0.4.3] - 2026-02-23
 ### Changed
 - Update dependencies.
@@ -161,6 +165,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Introduce new package. [#31163]
 
+[0.4.4]: https://github.com/Automattic/jetpack-boost-core/compare/v0.4.3...v0.4.4
 [0.4.3]: https://github.com/Automattic/jetpack-boost-core/compare/v0.4.2...v0.4.3
 [0.4.2]: https://github.com/Automattic/jetpack-boost-core/compare/v0.4.1...v0.4.2
 [0.4.1]: https://github.com/Automattic/jetpack-boost-core/compare/v0.4.0...v0.4.1